requirements.txt
python-telegram-bot==20.8
httpx==0.24.1
requests
flask
